+++ c4a-impl/src/main/ts/eu/xenit/care4alf/web/js/export.ts
@@ -1,7 +1,7 @@
 care4alf.controller('export', ($scope,$http: ng.IHttpService) => {
     $scope.form = {};
     $scope.form.query = "PATH:\"/app:company_home/cm:Projects_x0020_Home//*\" AND TYPE:\"cm:content\"";
-    $scope.form.columns = "cm:name,cm:creator,type,path";
+    $scope.form.columns = "cm:name,cm:creator,type,content.contentUrl,content.encoding,content.locale,content.mimetype,content.size,path";
     $scope.form.separator = ",";
     $scope.form.nullValue = "null";
     $scope.form.documentName = "export.csv";